Renaming the test file.
