+++ src/package.json
@@ -45,7 +45,7 @@
     "start:dynamodblocal": "cd tests/contract; ./scripts/setup-dynamodb-local.sh",
     "start:contract": "npm-run-all -s start:dynamodblocal wait-start:contract",
     "wait-start:contract": "wait-on tcp:127.0.0.1:8000 && cd tests/contract && ts-node ./app.ts",
-    "test:contract": "wait-on tcp:127.0.0.1:3000; cd tests/contract; ts-node ./test.ts",
+    "test:contract": "wait-on tcp:127.0.0.1:3000; cd tests/contract; ts-node ./*.ts",
     "test:contract:ci": "cross-env USE_MOCKED=true AWS_ACCESS_KEY_ID=dummy AWS_SECRET_ACCESS_KEY=dummy AWS_PAGER=\"\" npm-run-all -p -r start:contract test:contract"
   },
   "devDependencies": {
